Refactor `variables` GC into its own file.
