+++ lib/storage/exchanges.js
@@ -12,6 +12,9 @@ import assert from 'assert-plus';
 import {logger} from '../logger.js';
 import {serializeError} from 'serialize-error';
 
+// ensure externalized `variables` garbage collector runs
+import './variablesGarbageCollector.js';
+
 const {util: {BedrockError}} = bedrock;
 
 /* Note: Exchanges have default TTLs of 15 minutes and are always in one of
