+++ lib/storage/variables.js
@@ -7,10 +7,8 @@ import {
   EXCHANGE_EXPIRY_GRACE_PERIOD, EXCHANGE_TTL_DEFAULT
 } from '../constants.js';
 import {createHash} from 'node:crypto';
-import {logger} from '../logger.js';
 import {PassThrough} from 'node:stream';
 import {pipeline} from 'node:stream/promises';
-import {rangeDelay} from 'delay';
 import {buffer as readIntoBuffer} from 'node:stream/consumers';
 
 const {util: {BedrockError}} = bedrock;
@@ -41,7 +39,7 @@ will require different HTTP body parser setup). */
 const VARIABLES_EXPIRY_GRACE_PERIOD = 1000 * 60 * 60;
 
 // for gridfs storage of large exchange `variables`
-const VARIABLES_STORAGE = {
+export const VARIABLES_STORAGE = {
   name: 'vc-exchange-variables',
   bucket: null
 };
@@ -55,15 +53,6 @@ const VARIABLES_SIZE_LIMIT = 1024 * 1024 * 16;
 // used to determine whether `variables` can be stored parsed as BSON or must
 // be converted to JSON
 const MONGODB_ILLEGAL_KEY_CHAR_REGEX = /[%$.]/;
-
-// state for running a garbage collector for expired externalized `variables`
-const VARIABLES_GARBAGE_COLLECTOR = {
-  // used to abort variables garbage collector
-  abortController: new AbortController(),
-  // a Promise that resolves after the `variables` garbage collector has
-  // shutdown cleanly after receiving an abort signal
-  shutdownPromise: null
-};
 
 bedrock.events.on('bedrock-mongodb.ready', async () => {
   await database.openCollections([`${VARIABLES_STORAGE.name}.files`]);
@@ -97,27 +86,6 @@ bedrock.events.on('bedrock-mongodb.ready', async () => {
       unique: false
     }
   }]);
-});
-
-bedrock.events.on('bedrock.ready', () => {
-  // start the `variables` garbage collector, which runs continuously
-  VARIABLES_GARBAGE_COLLECTOR.shutdownPromise =
-    _startVariablesGarbageCollector();
-});
-
-bedrock.events.on('bedrock.exit', async () => {
-  try {
-    // abort variables garbage collector
-    VARIABLES_GARBAGE_COLLECTOR.abortController.abort();
-    logger.debug(
-      'Sent abort signal to "variables" garbage collector, ' +
-      'waiting for shutdown...');
-    await VARIABLES_GARBAGE_COLLECTOR.shutdownPromise;
-    logger.debug('"Variables" garbage collector shutdown was successful.');
-  } catch(error) {
-    logger.error(
-      'Error during "variables" garbage collector shutdown.', {error});
-  }
 });
 
 export async function encodeVariables({workflowId, exchange, meta}) {
@@ -236,31 +204,6 @@ export async function decodeVariables({workflowId, record}) {
   return exchange;
 }
 
-async function _deleteExpiredVariables({signal}) {
-  signal.throwIfAborted();
-
-  // delete all files found (limit=1000)
-  const {bucket} = VARIABLES_STORAGE;
-  const now = new Date(Date.now() + 86400 * 1000 * 365);
-  const projection = {_id: 1};
-  const cursor = bucket.find({
-    'metadata.expires': {$lte: now}
-  }, {projection}).limit(1000);
-  for await (const {_id} of cursor) {
-    try {
-      signal.throwIfAborted();
-      await bucket.delete(_id);
-    } catch(e) {
-      // ignore file not found errors and throw all others; note: there is
-      // currently not a better way to check for a `not found` error than to
-      // check the message text
-      if(!e.message.includes('not found')) {
-        throw e;
-      }
-    }
-  }
-}
-
 function _hasIllegalMongoDBKeyChar(value) {
   if(Array.isArray(value)) {
     for(const e of value) {
@@ -293,22 +236,4 @@ function _multibaseMultihashSha256(buffer) {
 
   // return as multibase-base64url-encoded value
   return 'u' + Buffer.from(mh).toString('base64url');
-}
-
-async function _startVariablesGarbageCollector() {
-  const {variablesGarbageCollector: {interval}} = bedrock.config['vc-workflow'];
-  const {signal} = VARIABLES_GARBAGE_COLLECTOR.abortController;
-  while(!signal.aborted) {
-    try {
-      // collect expired externalized exchange "variables" then delay for
-      // `interval` plus some fuzzing (up to 1 minute) to spread load
-      await _deleteExpiredVariables({signal});
-      await rangeDelay(interval, interval + 60000, {signal});
-    } catch(e) {
-      if(e.name === 'AbortError') {
-        break;
-      }
-      logger.error('Error in "variables" garbage collector job.', {error: e});
-    }
-  }
 }
